Fixed bug and implemented first test condition function for acceptance
Bug was that the code was recognising the target transaction in it's own block as a conflicting transaction when it was searching through the transactions in other blocks.

First test case function for acceptance was a function that incorporated the logic to see if the conflicting transactions in the anitcone set were preceded by the block of the target transaction. Was implemented to solve problems with 'break' - if would stop the simulation if algorithm 2 was implemented exactly as presented in the SPECTRE paper